// packages/test/src/types.ts
// TODO: Get these from OpenAPI spec directly
export type ApiVersion = "V2.0" | "V2.1" | "V2.2" | "V2.3" | "V3.0";

export enum EventTypesV2 {
  CREATED = "org.wbcsd.pathfinder.ProductFootprintRequest.Created.v1",
  FULFILLED = "org.wbcsd.pathfinder.ProductFootprintRequest.Fulfilled.v1",
  REJECTED = "org.wbcsd.pathfinder.ProductFootprintRequest.Rejected.v1",
  PUBLISHED = "org.wbcsd.pathfinder.ProductFootprint.Published.v1",
}

export enum EventTypesV3 {
  CREATED = "org.wbcsd.pact.ProductFootprint.RequestCreatedEvent.3",
  FULFILLED = "org.wbcsd.pact.ProductFootprint.RequestFulfilledEvent.3",
  REJECTED = "org.wbcsd.pact.ProductFootprint.RequestRejectedEvent.3",
  PUBLISHED = "org.wbcsd.pact.ProductFootprint.PublishedEvent.3",
}

// Constants for test run status
export enum TestRunStatus {
  PASS = "PASS",
  FAIL = "FAIL",
  PENDING = "PENDING",
}

// Constants for test result status
export enum TestCaseResultStatus {
  PENDING = "PENDING",
  SUCCESS = "SUCCESS",
  FAILURE = "FAILURE",
}

export type LogEntry =
  | { message: string }
  | { warning: string }
  | { error: string }
  | { request: { method: string; url: string; headers: Record<string, string>; data?: any; text?: string } }
  | { response: { statusCode: number; headers: Record<string, string>; data?: any; text?: string } };

export interface TestResult {
  name: string;
  status: TestCaseResultStatus;
  errorMessage?: string;
  apiResponse?: string;
  mandatory: boolean;
  testKey: string;
  curlRequest?: string;
  documentationUrl?: string;
  log?: LogEntry[];
}

export interface TestRunStartParams {
  baseUrl: string;
  version: ApiVersion;
  clientId: string;
  clientSecret: string;
  organizationName: string;
  adminEmail: string;
  adminName: string;
  customAuthBaseUrl?: string;
  scope?: string;
  audience?: string;
  resource?: string;
  testCaseNumbers?: number[];
}

export interface TestRun {
  testRunId: string;
  organizationName: string;
  adminEmail: string;
  adminName: string;
  timestamp: string;
  status: TestRunStatus;
  techSpecVersion: string;
  passingPercentage?: number;
  data: unknown; // jsonb, flexible field to store additional data like productIds;
}

export interface TestRunWithResults extends TestRun {
  results: TestResult[];
}

export interface PagingParameters {
  query?: string;
  page?: string;
  pageSize?: string;
}

/**
 * Interface representing a storage mechanism for managing test runs,
 * test results, and related data.
 * This can be implemented using different storage backends, like a database,
 * but could also be output to files or the console for simpler use cases.
 * For a one-off test run, a simple console or file-based implementation might suffice.
 * For persistent storage and historical data, a database implementation is recommended.
 */
export interface TestStorage {
  saveTestRun(testRun: TestRun): Promise<void>;
  getTestRun(testRunId: string): Promise<TestRun>;
  updateTestRunStatus(testRunId: string): Promise<void>;
  saveTestCaseResults(
    testRunId: string,
    testResults: TestResult[],
    overwriteExisting: boolean
  ): Promise<void>;
  getTestRunWithResults(testRunId: string): Promise<TestRunWithResults>;
  listTestRuns(
    paging: PagingParameters,
    adminEmail?: string
  ): Promise<TestRun[]>;
}